+++ src/Commands.cpp
@@ -259,6 +259,9 @@ static SeqStrings gCommandNames =
     "CmdChangeTheme\0"
     "CmdTranslateSelection\0"
     "CmdFavoriteShowInTab\0"
+    "CmdTocExpandToLevel1\0"
+    "CmdTocExpandToLevel2\0"
+    "CmdTocExpandToLevel3\0"
     "CmdNone\0"
     "\0";
 
@@ -511,6 +514,9 @@ static i32 gCommandIds[] = {
     CmdChangeTheme,
     CmdTranslateSelection,
     CmdFavoriteShowInTab,
+    CmdTocExpandToLevel1,
+    CmdTocExpandToLevel2,
+    CmdTocExpandToLevel3,
     CmdNone,
 };
 
@@ -763,6 +769,9 @@ SeqStrings gCommandDescriptions =
     "Change Theme...\0"
     "Translate Selection...\0"
     "Show Favorites in Tab\0"
+    "Bookmarks: Expand to Level 1\0"
+    "Bookmarks: Expand to Level 2\0"
+    "Bookmarks: Expand to Level 3\0"
     "Do nothing\0"
     "\0";
 // clang-format on